+++ frontend/src/App.tsx
@@ -11,7 +11,6 @@ import { Toaster } from "sonner";
 import UpdateCheckController from "./components/modals/UpdateCheckController";
 import VersionInfoController from "./components/modals/VersionInfoController";
 import AttachmentMaintenanceController from "./components/modals/AttachmentMaintenanceController";
-import MarkdownSettingsController from "./components/modals/MarkdownSettingsController";
 
 export default function App() {
   return (
@@ -32,7 +31,6 @@ export default function App() {
           <UpdateCheckController />
           <VersionInfoController />
           <AttachmentMaintenanceController />
-          <MarkdownSettingsController />
           <Toaster
             position="bottom-center"
             expand={false}